fix: fixes tests

+++ test/helpers/emptyDirectory.js
@@ -6,9 +6,11 @@ const emptyDirectory = directoryPath => new Promise((resolve, reject) => {
     fs.readdir(directoryPath, (error, files) => {
         if (error) reject(error);
         for (const file of files) {
-            fs.unlink(path.join(directoryPath, file), error => {
-                if (error) reject(error);
-            });
+            if (file !== '.gitignore') {
+                fs.unlink(path.join(directoryPath, file), error => {
+                    if (error) reject(error);
+                });
+            }
         }
         resolve();
     });
